@@ -1,74 +1,35 @@
-import Boss from '../_models/boss.js';
-import Worker from '../_models/worker.js';
+import {User} from '../_models/user.js';
+import dotenv from "dotenv"
+import jwt from "jsonwebtoken";
 
-export const newBoss =  async (req,res)=>{
-	try{
-		const { name, password, location } = req.body;
-		const newBoss = await new Boss({
-			name,
-			password,
-			location
-		});
-		const savedBoss = await newBoss.save();
-		res.status(200).json({message: savedBoss._id,
-							  success:true
-	});
-	}catch(err){
-		res.status(400).json(false);
-	}
-}
+dotenv.config();
 
-export const newWorker =  async (req,res)=>{
+export const register =  async (req,res)=>{
 	try{
-		const { name, password, location, bossID, skillset } = req.body;
-		const newWorker = await new Worker({
+		const {email,name,password,location} = req.body;
+		
+		if (!(email && name && password && location)) {
+			res.status(400).send("All params required");
+		}
+		const newUser = await new User({
+			email,
 			name,
 			password,
 			location,
-			bossID,
-			skillset
+			skills:[],
+			bosses:[],
+			workers:[]
 		});
-		const savedWorker = await newWorker.save();
-		res.status(200).json({message:`Your userId is ${savedWorker._id}`,
-							  success:true
-		});
+		const savedUser = await newUser.save();
+		const token = jwt.sign(
+			{ user_id: savedUser._id, email },
+			process.env.SECRET,
+			{
+			  expiresIn: "2h",
+			}
+		);
+		res.status(200).json({token});
 	}catch(err){
-		console.log(err);
-	}
-}
-export const loginFunc = async (req,res) => {
-	try{
-		const { bossID, password } = req.body;
-		try{
-			const bossData = await Boss.findById(bossID);
-			const bossName = bossData.name
-			const idPassword = bossData.password;
-			const correct =  (password === idPassword);
-			res.json({success :correct,
-					  name :bossName
-			})
-		}catch(err){
-			res.json(false)}
-		
-		// console.log(pass)
-	}catch(err){
-		console.log(err)
-	}
-}
-export const loginFuncWorker = async(req,res) => {
-	try{
-		const {workerID, password, bossID} = req.body;
-		try{
-			const workerData = await Worker.findById(workerID);
-			const passwordCorrect = workerData.password === password;
-			const bossIDCorrect = workerData.bossID === bossID;
-			console.log(workerData.bossID);
-			passwordCorrect && bossIDCorrect ?
-				res.json(true):
-					res.json(false)
-		}catch(err){
-			res.json(false)}
-	}catch(err){
-		console.log(err)
+		res.status(400);
 	}
 }
