@@ -1,12 +1,8 @@
 import express from 'express';
-import { newBoss, newWorker, loginFunc, loginFuncWorker } from '../_controllers/dbControllers.js';
+import { register } from '../_controllers/dbControllers.js';
 
-export const bossRouter = express.Router();
+export const userRouter = express.Router();
 
-bossRouter.route('/post').post(newBoss);
-bossRouter.route('/login').post(loginFunc);
+userRouter.route('/register').post(register);
+// userRouter.route('/login').post(loginFuncBoss);
 
-export const workerRouter = express.Router();
-
-workerRouter.route('/post').post(newWorker);
-workerRouter.route('/login').post(loginFuncWorker);
