@@ -2,9 +2,9 @@ import React from 'react';
 
 const MainScreenWorker = ()=>{
     return(
-        <>
-        
-        </>
-    )
+		<div>
+			
+		</div>
+    );
 }
 export default MainScreenWorker;
